move whisper logic to subscriptions
(cherry picked from commit 7f2a1d67a4a1ef929e3ac5705c4088e03a790d99)
