--- lib/anycable/rails/action_cable_ext/whisper_stream.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-require "action_cable"
-
-ActionCable::Connection::Subscriptions.prepend(Module.new do
-  def execute_command(data)
-    return whisper(data) if data["command"] == "whisper"
-
-    super
-  end
-
-  def whisper(data)
-    find(data).whisper(ActiveSupport::JSON.decode(data["data"]))
-  end
-end)
