+++ lib/anycable/rails/ext/signed_streams.rb
